Styling and interface fixes
Fixed an issue with the left CMP icon not aligning to the left on Safari. Fixed an issue with multi-select on the Training page where the import/export feature wouldn't work for all multi-selects. Selected team now clears when entering a new Tournament Mode game to fix an issue that sometimes prevented a full team from being selected.