Move count updates, custom group QOL
Centralized move count math in a static method of the Pokemon class. Move counts of "6, 6, 5" and "6, 5, 6" are now distinguished by a period sign and minus sign respectively. Alternate Fast Moves can now be selected in the rankings to view move counts for other movesets. Adding a new Pokemon to a custom group scrolls the list to the bottom.