Remove cup specific ranking iterations
The iteration value repeats the process of weighing matchups by the assigned weights and overall scores of the opponents.  Most rankings have operated on 1 weight cycle for a while, so cup specific iteration values aren't needed. (If this were to be reimplemented, it should be a property of the cup JSON.)